Add files via upload
This is the file with all the code. *Note: you'll have to install requests with pip3 install requests*
Also, It has a bug that I'm not sure about, so if you'd be so kind as to leave an issue, That would be great. Finally, you can have it convert anything by changing from *USD* to another code for the country.